Theme moved into extended controls. Skinny sysinfo look. Half way into Mat commds changes
