+++ EDDiscovery/EDDiscovery.csproj
@@ -497,7 +497,6 @@
       <DependentUpon>UserControlForm.cs</DependentUpon>
     </Compile>
     <Compile Include="HistoryList.cs" />
-    <Compile Include="Themeing\CustomColourTable.cs" />
     <Compile Include="DB\SavedRouteClass.cs" />
     <Compile Include="EliteDangerous\ISystem.cs" />
     <Compile Include="EliteDangerous\SystemClass.cs" />
@@ -506,7 +505,6 @@
     <Compile Include="DB\WantedSystemClass.cs" />
     <Compile Include="EDDConfig.cs" />
     <Compile Include="Themeing\Theme.cs" />
-    <Compile Include="Themeing\ToolStripRenderer.cs" />
     <Compile Include="EDSM\EDSMClass.cs" />
     <Compile Include="EDSM\EDSMSync.cs" />
     <Compile Include="EliteDangerous\EDCommander.cs" />
@@ -532,12 +530,6 @@
     </Compile>
     <Compile Include="Forms\AboutForm.Designer.cs">
       <DependentUpon>AboutForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Forms\ThemeEditor.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="Forms\ThemeEditor.Designer.cs">
-      <DependentUpon>ThemeEditor.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\MoveToCommander.cs">
       <SubType>Form</SubType>
@@ -1145,9 +1137,6 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\RecordStep.resx">
       <DependentUpon>RecordStep.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Forms\ThemeEditor.resx">
-      <DependentUpon>ThemeEditor.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="ImageHandler.resx">
       <DependentUpon>ImageHandler.cs</DependentUpon>
