+++ ExtendedControls/ExtendedControls.csproj
@@ -35,6 +35,7 @@
     <Reference Include="System.Core" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Windows.Forms" />
+    <Reference Include="System.Windows.Forms.DataVisualization" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -60,6 +61,13 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="BitMapHelpers.cs" />
+    <Compile Include="ThemeStandardEditor.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ThemeStandardEditor.Designer.cs">
+      <DependentUpon>ThemeStandardEditor.cs</DependentUpon>
+    </Compile>
+    <Compile Include="ThemeToolStrip.cs" />
     <Compile Include="CustomDateTimePicker.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -139,9 +147,8 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="Theme.cs" />
-    <Compile Include="ToolStripComboBoxCustom.cs">
-      <SubType>Component</SubType>
-    </Compile>
+    <Compile Include="ThemeStandard.cs" />
+    <Compile Include="ToolStripComboBoxCustom.cs" />
     <Compile Include="UpDown.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -174,6 +181,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="TabStrip.resx">
       <DependentUpon>TabStrip.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="ThemeStandardEditor.resx">
+      <DependentUpon>ThemeStandardEditor.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
