+++ ExtendedControls/ThemeStandardEditor.resx
